requirements.txt
django
gunicorn
django-heroku
Pillow
django-crispy-forms
whitenoise
dj-database-url
psycopg2-binary
disqus
api
docutils-stubs
django-markupfield
